+++ src/main/kotlin/au/gov/api/EventManager.kt
@@ -23,8 +23,6 @@ import com.fasterxml.jackson.databind.ObjectMapper
 @Component
 class EventManager{
 
-        @Value("\${spring.datasource.url}")
-        private var dbUrl: String? = null
 
         @Autowired
         private lateinit var dataSource: DataSource
@@ -90,19 +88,4 @@ class EventManager{
             }
         }
 
-        @Bean
-        @Throws(SQLException::class)
-        fun dataSource(): DataSource? {
-            if (dbUrl?.isEmpty() ?: true) {
-                return HikariDataSource()
-            } else {
-                val config = HikariConfig()
-                config.jdbcUrl = dbUrl
-                try {
-                    return HikariDataSource(config)
-                } catch (e: Exception) {
-                    return null
-                }
-            }
-        }
 }
